move perceptual loss to dedicated module
