+++ pystiche/loss/__init__.py
@@ -1,1 +1,2 @@
-from .loss import *
+from .multi_op import *
+from .perceptual import *
